New search for download now resets scroller position, some refactoring
