+++ AudicaMod/SongBrowser.csproj
@@ -240,6 +240,7 @@
     <Compile Include="csvorbis\Time0.cs" />
     <Compile Include="csvorbis\VorbisFile.cs" />
     <Compile Include="csvorbis\OggStream.cs" />
+    <Compile Include="src\SongDownloader.cs" />
     <Compile Include="src\UI\Buttons\SongSearchButton.cs" />
     <Compile Include="src\UI\SongLoadingManager.cs" />
     <Compile Include="src\UI\SongSearchScreen.cs" />
