+++ AudicaMod/src/SongDownloader.cs
@@ -1,0 +1,126 @@
+using System;
+using System.Collections;
+using System.Net;
+using MelonLoader;
+using UnityEngine;
+using MelonLoader.TinyJSON;
+using System.IO;
+using System.Media;
+using OggDecoder;
+
+namespace AudicaModding
+{
+    public static class SongDownloader
+    {
+        public static string      apiURL = "http://www.audica.wiki:5000/api/customsongs?pagesize=14";
+        public static APISongList songlist;
+        public static string      searchString = "";
+        public static bool        needRefresh = false;
+        public static int         page = 1;
+
+        private static IEnumerator PlayOggCoroutine(string oggFilename)
+        {
+            using (var file = new FileStream(oggFilename, FileMode.Open, FileAccess.Read))
+            {
+                var player = new SoundPlayer(new OggDecodeStream(file));
+                player.Play();
+                yield return new WaitForSeconds(10f);
+            }
+            yield return null;
+        }
+
+        public static void StartNewSongSearch()
+        {
+            page = 1;
+            StartNewPageSearch();
+        }
+
+        public static void StartNewPageSearch()
+        {
+            SongDownloaderUI.ResetScrollPosition();
+            MelonCoroutines.Start(StartSongSearchCoroutine(searchString, SongDownloaderUI.difficultyFilter.ToString(), page, false));
+        }
+
+        public static IEnumerator StartSongSearchCoroutine(string search, string difficulty = null, int page = 1, bool total = false)
+        {
+            string webSearch = search == null || search == "" ? "" : "&search=" + WebUtility.UrlEncode(search);
+            string webPage = page == 1 ? "" : "&page=" + page.ToString();
+            string webDifficulty = difficulty == "All" || difficulty == "" ? "" : "&" + difficulty.ToLower() + "=true";
+            string webCurated = SongDownloaderUI.curated ? "&curated=true" : "";
+            string webPlaycount = SongDownloaderUI.popularity ? "&sort=leaderboards" : "";
+            string concatURL = !total ? apiURL + webSearch + webDifficulty + webPage + webCurated + webPlaycount : "http://www.audica.wiki:5000/api/customsongs?pagesize=all";
+            WWW www = new WWW(concatURL);
+            yield return www;
+            songlist = JSON.Load(www.text).Make<APISongList>();
+            if (SongDownloaderUI.songItemPanel != null)
+            {
+                SongDownloaderUI.AddSongItems(SongDownloaderUI.songItemMenu, songlist);
+            }
+        }
+
+
+        public static IEnumerator DownloadSong(string downloadUrl)
+        {
+            string[] splitURL = downloadUrl.Split('/');
+            string audicaName = splitURL[splitURL.Length - 1];
+            string path = Application.streamingAssetsPath + "\\HmxAudioAssets\\songs\\" + audicaName;
+            string downloadPath = SongBrowser.downloadsDirectory + "\\" + audicaName;
+            if (!File.Exists(path) && !File.Exists(downloadPath) && !File.Exists(downloadPath))
+            {
+                WWW www = new WWW(downloadUrl);
+                yield return www;
+                byte[] results = www.bytes;
+                File.WriteAllBytes(downloadPath, results);
+                needRefresh = true;
+            }
+            yield return null;
+        }
+
+        public static IEnumerator StreamPreviewSong(string url)
+        {
+            WWW www = new WWW(url);
+            yield return www;
+            if (www.isDone)
+            {
+                Stream stream = new MemoryStream(www.bytes);
+                var player = new SoundPlayer(new OggDecodeStream(stream));
+                //player.LoadAsync();
+                yield return new WaitForSeconds(0.2f);
+                player.Play();
+                yield return new WaitForSeconds(15f);
+            }
+
+            yield return null;
+        }
+
+        public static void NextPage()
+        {
+            if (page > songlist.total_pages)
+                page = songlist.total_pages;
+            else if (page < 1)
+                page = 1;
+            else
+                page++;
+        }
+        public static void PreviousPage()
+        {
+            if (page == 1) return;
+            if (page > songlist.total_pages)
+                page = songlist.total_pages;
+            else if (page < 1)
+                page = 1;
+            else
+                page--;
+        }
+    }
+}
+
+[Serializable]
+public class APISongList
+{
+    public int total_pages;
+    public int song_count;
+    public Song[] songs;
+    public int pagesize;
+    public int page;
+}
