+++ AudicaMod/src/UI/SongDownloaderUI.cs
@@ -36,7 +36,7 @@ namespace AudicaModding
 
 		public static void GoToWebSearchPage()
 		{
-			SongBrowser.page = 1;
+			SongDownloader.page = 1;
 			if (songItemPanel == null)
 			{
 				secondaryPanel = GameObject.Instantiate(GameObject.Find("ShellPage_Settings"));
@@ -55,6 +55,11 @@ namespace AudicaModding
 			primaryMenu.screenTitle.text = "Filters";
 			SongBrowser.lastSongCount = SongBrowser.newSongCount; //User has seen new songs
 			SongBrowser.SavePrivateConfig();
+		}
+
+		public static void ResetScrollPosition()
+        {
+			songItemMenu?.scrollable.SnapTo(0);
 		}
 		
 		private static void SpawnSecondaryPanel(GameObject secondaryPanel)
@@ -94,8 +99,7 @@ namespace AudicaModding
 					difficultyFilter++;
 					if ((int)difficultyFilter > 4) difficultyFilter = 0;
 					difficultyToggle.label.text = difficultyFilter.ToString();
-					SongBrowser.page = 1;
-					SongBrowser.StartSongSearch();
+					SongDownloader.StartNewSongSearch();
 				}),
 				null,
 				"Filters the search to the selected difficulty");
@@ -118,8 +122,7 @@ namespace AudicaModding
 						curated = true;
 
 					curatedToggle.label.text = "Curated only: " + curated.ToString();
-					SongBrowser.page = 1;
-					SongBrowser.StartSongSearch();
+					SongDownloader.StartNewSongSearch();
 				}),
 				null,
 				"Filters the search to curated maps only");
@@ -160,8 +163,7 @@ namespace AudicaModding
 						popularity = true;
 
 					popularityToggle.label.text = "Sort by playcount: " + popularity.ToString();
-					SongBrowser.page = 1;
-					SongBrowser.StartSongSearch();
+					SongDownloader.StartNewSongSearch();
 				}),
 				null,
 				"Sorts downloads by leaderboard scores rather than date.");
@@ -177,7 +179,7 @@ namespace AudicaModding
 		{
 			foreach (var song in activeSongList.songs)
 			{
-				MelonCoroutines.Start(SongBrowser.DownloadSong(song.download_url));
+				MelonCoroutines.Start(SongDownloader.DownloadSong(song.download_url));
 			}
 		}
 
@@ -185,7 +187,7 @@ namespace AudicaModding
 		{
 			CleanUpPage(optionsMenu);
 			activeSongList = songlist;
-			optionsMenu.screenTitle.text = "Displaying page " + SongBrowser.page.ToString() + " out of " + songlist.total_pages.ToString();
+			optionsMenu.screenTitle.text = "Displaying page " + SongDownloader.page.ToString() + " out of " + songlist.total_pages.ToString();
 
 			var pageHeader = optionsMenu.AddHeader(0, "Listing " + songlist.song_count.ToString() + " songs");
 			optionsMenu.scrollable.AddRow(pageHeader.gameObject);
@@ -205,14 +207,14 @@ namespace AudicaModding
 			var row = new Il2CppSystem.Collections.Generic.List<GameObject>();
 			var previousPage = optionsMenu.AddButton(0,
 				"Previous Page",
-				new Action(() => { SongBrowser.PreviousPage(); SongBrowser.StartSongSearch(); optionsMenu.scrollable.SnapTo(0); }),
+				new Action(() => { SongDownloader.PreviousPage(); SongDownloader.StartNewPageSearch(); optionsMenu.scrollable.SnapTo(0); }),
 				null,
 				null);
 			row.Add(previousPage.gameObject);
 
 			var nextPage = optionsMenu.AddButton(1,
 				"Next Page",
-				new Action(() => { SongBrowser.NextPage(); SongBrowser.StartSongSearch(); optionsMenu.scrollable.SnapTo(0); }),
+				new Action(() => { SongDownloader.NextPage(); SongDownloader.StartNewPageSearch(); optionsMenu.scrollable.SnapTo(0); }),
 				null,
 				null);
 			row.Add(nextPage.gameObject);
@@ -250,7 +252,7 @@ namespace AudicaModding
 
             var downloadButton = optionsMenu.AddButton(0,
 				"Download" + SongBrowser.GetDifficultyString(songd),
-				new Action(() => { MelonCoroutines.Start(SongBrowser.DownloadSong(song.download_url)); TMP.text = "Added song to download queue!"; }),
+				new Action(() => { MelonCoroutines.Start(SongDownloader.DownloadSong(song.download_url)); TMP.text = "Added song to download queue!"; }),
 				null,
 				null);
 			downloadButton.button.destroyOnShot = true;
@@ -259,7 +261,7 @@ namespace AudicaModding
 
 			var previewButton = optionsMenu.AddButton(1,
 				"Preview",
-				new Action(() => { MelonCoroutines.Start(SongBrowser.StreamPreviewSong(song.preview_url)); }),
+				new Action(() => { MelonCoroutines.Start(SongDownloader.StreamPreviewSong(song.preview_url)); }),
 				null,
 				null);
 			row.Add(previewButton.gameObject);
@@ -288,7 +290,7 @@ namespace AudicaModding
 		{
 			yield return new WaitForSeconds(0.05f);
 			SetupSecondaryPanel(panel);
-			AddSongItems(panel.GetComponentInChildren<OptionsMenu>(), SongBrowser.songlist);
+			AddSongItems(panel.GetComponentInChildren<OptionsMenu>(), SongDownloader.songlist);
 		}
 	}
 }
